Authenticate - fix inherited properties not initialised
When used in combination with other modules in this distribution,
and depending on the order in which they were loaded, the inherited
`Request` properties were not being initialised for the
`WebService::HMRC::Authenticate` module.

This patch changes `Authenticate` from being a compile-time `use`
dependecy, to a run-time `require` dependency for the `Request` module,
which is loaded only when necessary.

A test is added for the previous failure